third_party/blink/web_tests/external/wpt/html/syntax/parsing/html5lib_tests7_run_type=uri-expected.txt

This is a testharness.js-based test.
[FAIL] html5lib_tests7.html 84e2152c284f4dfee7d8d12846c08b2c025578a6
  assert_equals: expected "#document\\n| <!DOCTYPE html>\\n| <html>\\n|   <head>\\n|   <body>\\n|     <select>\\n|     <input>\\n|     \\"X\\"" but got "#document\\n| <!DOCTYPE html>\\n| <html>\\n|   <head>\\n|   <body>\\n|     <select>\\n|       <input>\\n|       \\"X\\""
Harness: the test ran to completion.
